# apps/backend/src/main/resources/swagger/api.yml
...
openapi: '3.0.1'
info:
  title: 'backend'
  version: 0.0.1
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://localhost:8080/api
    description: Development server with TLS Profile
paths:
  /v1/grota-financiamentos/auth/register:
    post:
      tags:
        - Grota Financiamentos Auth
      operationId: registerDealer
      summary: Cadastrar Lojista
      description: Cadastra um Lojista no banco de dados
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '201':
          description: Lojista cadastrada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealerAdminRegisterResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
...
components:
  schemas:
    AuthRegisterRequest:
      type: object
      required:
        - fullName
        - phone
        - enterprise
        - password
      properties:
        fullName:
          type: string
        phone:
          type: string
        enterprise:
          type: string
        password:
          type: string
        adminRegistration:
          type: boolean
...
    DealerAdminRegisterResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        razaoSocial:
          type: string
        cnpj:
          type: string
        referenceCode:
          type: string
        phone:
          type: string
        enterprise:
          type: string
        logoUrl:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - ATIVO
            - PENDENTE
            - INATIVO
        createdAt:
          type: string
          format: date-time
...
  responses:
    Problem:
      description: error occurred - see status code and problem object for more information.
      content:
        application/problem+json:
          schema:
            type: object
            properties:
              type:
                type: string
                format: uri
                description: |
                  An absolute URI that identifies the problem type. When dereferenced,
                  it SHOULD provide human-readable documentation for the problem type
                  (e.g., using HTML).
                default: 'about:blank'
                example: 'https://www.jhipster.tech/problem/problem-with-message'
              title:
                type: string
                description: |
                  A short, summary of the problem type. Written in english and readable
                  for engineers (usually not suited for non technical stakeholders and
                  not localized); example: Service Unavailable
              status:
                type: integer
                format: int32
                description: |
                  The HTTP status code generated by the origin server for this occurrence
                  of the problem.
                minimum: 100
                maximum: 600
                exclusiveMaximum: true
                example: 503
              detail:
                type: string
                description: |
                  A human readable explanation specific to this occurrence of the
                  problem.
                example: Connection to database timed out
              instance:
                type: string
                format: uri
                description: |
                  An absolute URI that identifies the specific occurrence of the problem.
                  It may or may not yield further information if dereferenced.
  securitySchemes:
    oauth:
      type: oauth2
      description: OAuth2 authentication with KeyCloak
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:9080/realms/jhipster/protocol/openid-connect/auth
          tokenUrl: http://localhost:9080/realms/jhipster/protocol/openid-connect/token
          scopes:
            jhipster: Jhipster specific claims
            email: Email claims
            profile: Profile claims
    openId:
      type: openIdConnect
      description: OpenID Connect authentication with KeyCloak
      openIdConnectUrl: http://localhost:9080/realms/jhipster/.well-known/openid-configuration
security:
  - oauth: [jhipster, email, profile]
  - openId: [jhipster, email, profile]